+++ module/test-dialog.js
@@ -22,6 +22,7 @@ const DEFAULT_TEST = {
   targetsModifier: 0,
   // attack-options
   calledShotModifier: 0,
+  concentratingModifier: 0,
   vitalAreaDamageModifier: false,
   burstModifier: 0,
   allOutFlag: false,
@@ -121,7 +122,10 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
       this.test.actorType ??= actor.type;
 
       const item = this.test.itemId ? actor.items.get(this.test.itemId) : null;
-      if (item) this.test.trademark = item.system.traits.has('trademark');
+      if (item) {
+        this.test.trademark = item.system.traits.has('trademark');
+        this.test.requiresConcentration = item.system.requiresConcentration;
+      }
     }
     // Ensure all relevant fields are Number
     for (const key of Object.keys(DEFAULT_TEST))
@@ -158,6 +162,15 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
     context.test.stymiedModifier = myActor.statusModifiers.stymied;
     context.test.darknessModifier = myActor.statusModifiers.darkness;
     context.test.waitingModifier = myActor.statusModifiers.waiting;
+    // TODO: only if the skill is affected by concentration:
+    // (Concentration check, any power skill tests)
+    // alteration, apportation, conjuration, divination, faith, psionics
+    // (willpower/spirit if Concentration check)
+    if (context.test.isConcentrationCheck ||
+      CONFIG.torgeternity.concentrationSkills.includes(context.test.skillName)) {
+      context.test.concentratingModifier = myActor.statusModifiers.concentrating;
+    }
+    context.test.requiresConcentration = this.test.itemId && myActor.items.get(this.test.itemId)?.requiresConcentration;
 
     // Set Modifiers for Vehicles
     if (this.test.testType === 'chase') {
@@ -202,6 +215,7 @@ export class TestDialog extends HandlebarsApplicationMixin(ApplicationV2) {
         context.test?.stymiedModifier ||
         context.test?.darknessModifier ||
         context.test?.waitingModifier ||
+        context.test?.concentratingModifier ||
         context.test?.sizeModifier ||
         context.test?.vulnerableModifier ||
         context.test?.speedModifier ||
@@ -340,6 +354,7 @@ export function oneTestTarget(token, applySize) {
       skills: actor.system.skills,
       attributes: actor.system.attributes,
       vulnerableModifier: actor.statusModifiers.vulnerable,
+      isConcentrating: actor.isConcentrating,
       defenses: {
         dodge: actor.defenses.dodge.value,
         unarmedCombat: actor.defenses.unarmedCombat.value,